+++ forui/example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 
 import 'package:forui/forui.dart';
 import 'package:forui_example/sandbox.dart';
@@ -8,8 +7,6 @@ import 'package:wakelock_plus/wakelock_plus.dart';
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   WakelockPlus.enable();
-
-  timeDilation = 1;
 
   runApp(const Application());
 }
@@ -47,7 +44,7 @@ class _ApplicationState extends State<Application> with SingleTickerProviderStat
     home: Builder(
       builder: (context) {
         return FScaffold(
-          header: const FHeader(title: Text('Example')),
+          header: const FHeader(title: Text('Example (B)')),
           footer: FBottomNavigationBar(
             index: index,
             onChange: (index) => setState(() => this.index = index),