+++ forui/test/src/widgets/header/nested_header_golden_test.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/widgets.dart';
 
 import 'package:flutter_test/flutter_test.dart';
@@ -47,6 +49,41 @@ void main() {
       );
 
       await expectLater(find.byType(TestScaffold), matchesGoldenFile('header/nested/${theme.name}.png'));
+    });
+
+    testWidgets('${theme.name} glassmorphic', (tester) async {
+      await tester.pumpWidget(
+        TestScaffold(
+          theme: theme.data,
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Stack(
+                children: [
+                  const Text('Lorem ipsum dolor sit amet, consectetur adipiscing elit.'),
+                  FHeader.nested(
+                    style: theme.data.headerStyles.nestedStyle.copyWith(
+                      backgroundFilter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
+                      decoration: BoxDecoration(color: theme.data.colors.background.withValues(alpha: 0.5)),
+                    ),
+                    title: const Text('Title'),
+                    prefixes: [
+                      FHeaderAction.back(onPress: () {}),
+                      const FHeaderAction(icon: Icon(FIcons.alarmClock), onPress: null),
+                    ],
+                    suffixes: [
+                      FHeaderAction(icon: const Icon(FIcons.plus), onPress: () {}),
+                      FHeaderAction.x(onPress: () {}),
+                    ],
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ),
+      );
+
+      await expectLater(find.byType(TestScaffold), matchesGoldenFile('header/nested/${theme.name}-glassmorphic.png'));
     });
 
     testWidgets('${theme.name} with no FNestedHeader actions', (tester) async {